requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
numpy>=1.21.0
scikit-learn>=1.3.0

# AI and embeddings (FREE resources)
openai>=1.0.0
chromadb>=0.4.0
python-dotenv>=1.0.0

# Text processing
nltk==3.8.1
textblob==0.17.1

# Visualization and analysis  
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0